Add vip to forward public traffic from hosts
Use 85.91.50.254 as a vip for cumulus to receive traffic that will be forwarded
directly to public networks. This will effectively make a subnet 85.91.50.240/28
for our hosts that need a static public ip. Virtual services deployed on kube
should use the remaining subnets under 85.91.50.0/24 to avoid ip collisions and
work as usual.