Adding the fallback default route for PROD / filtering default route
The fallback route will have metric 254. It makes this route only used
when the NAT boxes are not available ( if we are rebooting / rebuilding
the whole environment). As soon as the NAT boxes become  available they will
provide default routes with metric 200 (learned via iBGP), which will
have priority over this static route (254).

Also filtering the default route learned from merit house as they are
not in use.