Bug (new behaviour) introduced on Cumulus 4.3.0
Fixing the problems with MLAG introduced in 4.3.0:
https://docs.nvidia.com/networking-ethernet-software/cumulus-linux-43/Whats-New/rn/
https://support.mellanox.com/s/case/5001T00001YdbYeQAJ
After upgrading to Cumulus Linux, MLAG ports might remain down with clagctl and net show clag reporting bridge-priority-mismatch.
To work around this issue, run the sudo ifreload -a command on both peers, or configure bridge-bridgeprio to be the same value as mstpcl-treeprio on the bridge interface in the /etc/network/interfaces file, then run sudo ifreload -a.